Rename middlewares.
