+++ src/Middleware/AuthenticationMiddleware.php
@@ -1,136 +1,149 @@
 <?php
-/**
- * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
- * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- *
- * Licensed under The MIT License
- * For full copyright and license information, please see the LICENSE.txt
- * Redistributions of files must retain the above copyright notice.
- *
- * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- * @link          https://cakephp.org CakePHP(tm) Project
- * @since         1.0.0
- * @license       https://opensource.org/licenses/mit-license.php MIT License
- */
 namespace Authentication\Middleware;
 
+use Authentication\AuthenticationService;
 use Authentication\AuthenticationServiceInterface;
-use Authentication\AuthenticationServiceProviderInterface;
-use Authentication\Authenticator\UnauthenticatedException;
-use Authentication\Authenticator\UnauthorizedException;
-use Cake\Core\InstanceConfigTrait;
+use Authentication\Authenticator\AuthenticatorCollection;
+use Authentication\Authenticator\AuthenticatorCollectionInterface;
+use Authentication\Authenticator\PersistenceInterface;
+use Authentication\Authenticator\ResultInterface;
+use Authentication\Authenticator\StatelessInterface;
 use InvalidArgumentException;
+use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use RuntimeException;
-use Zend\Diactoros\Response\EmitterInterface;
-use Zend\Diactoros\Stream;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+use Symfony\Component\Translation\Exception\RuntimeException;
 
 /**
- * Authentication Middleware
+ * PSR 15 Authenticator Middleware
  */
-class AuthenticationMiddleware
+class AuthenticationMiddleware implements MiddlewareInterface
 {
-    use InstanceConfigTrait;
+    /**
+     * Response factory
+     *
+     * @var null|\Psr\Http\Message\ResponseFactoryInterface
+     */
+    protected $responseFactory;
 
     /**
-     * Configuration options
+     * Authentication failure redirect URL
      *
-     * - `identityAttribute` - The request attribute to store the identity in.
-     * - `name` the application hook method to call. Will be prefixed with `authentication`
-     * - `unauthenticatedRedirect` - The URL to redirect unauthenticated errors to. See
-     *    AuthenticationComponent::allowUnauthenticated()
-     * - `queryParam` - Set to true to have unauthenticated redirects contain a `redirect` query string
-     *   parameter with the previously blocked URL.
+     * @var string|callable
      */
-    protected $defaultConfig = [
-        'identityAttribute' => 'identity',
-        'name' => null,
-        'unauthenticatedRedirect' => null,
-        'queryParam' => null,
-    ];
+    protected $unauthorizedRedirectUrl;
 
     /**
-     * Authentication service or application instance.
+     * Successful login redirect URL
      *
-     * @var \Authentication\AuthenticationServiceInterface|\Authentication\AuthenticationServiceProviderInterface
+     * @var string|callable
      */
-    protected $subject;
+    protected $successRedirectUrl;
 
     /**
-     * Constructor
-     *
-     * @param \Authentication\AuthenticationServiceInterface|\Authentication\AuthenticationServiceProviderInterface $subject Authentication service or application instance.
-     * @param array|string $config Array of configuration settings or string with authentication service provider name.
-     * @throws \InvalidArgumentException When invalid subject has been passed.
+     * @var \Authentication\AuthenticationServiceInterface
      */
-    public function __construct($subject, $config = null)
+    protected $service;
+
+    /**
+     * AuthenticationPsr15Middleware constructor.
+     *
+     * @param null|\Psr\Http\Message\ResponseFactoryInterface $responseFactory
+     */
+    public function __construct(
+        AuthenticationServiceInterface $service,
+        ?ResponseFactoryInterface $responseFactory = null
+    ){
+        $this->service = $service;
+        $this->responseFactory = $responseFactory;
+    }
+
+    protected function setRedirect($redirectUrl, $type)
     {
-        if (is_string($config)) {
-            $config = ['name' => $config];
-        }
-        $this->setConfig($config);
-
-        if (!($subject instanceof AuthenticationServiceInterface) && !($subject instanceof AuthenticationServiceProviderInterface)) {
-            $expected = implode('` or `', [
-                AuthenticationServiceInterface::class,
-                AuthenticationServiceProviderInterface::class
-            ]);
-            $type = is_object($subject) ? get_class($subject) : gettype($subject);
-            $message = sprintf('Subject must be an instance of `%s`, `%s` given.', $expected, $type);
-
-            throw new InvalidArgumentException($message);
+        if (!is_string($redirectUrl) && !is_callable($redirectUrl)) {
+            throw new InvalidArgumentException('Redirect URL must be a string or callable');
         }
 
-        $this->subject = $subject;
+        if (!in_array($type, ['unauthorized', 'success'])) {
+            throw new InvalidArgumentException('Type must be failure or success');
+        }
+
+        $property = $type . 'RedirectUrl';
+        $this->{$property} = $redirectUrl;
+    }
+
+    public function setLoginRedirect($redirectUrl): self
+    {
+        return $this;
+
+        $this->setRedirect($redirectUrl, 'success');
+    }
+
+    public function setUnauhtorizedRedirect($redirectUrl): self
+    {
+        $this->setRedirect($redirectUrl, 'unauthorized');
+
+        return $this;
+    }
+
+    protected function getUnauthorizedRedirectResponse($request, $authResult, $authenticator): ResponseInterface
+    {
+        $response = $this->responseFactory->createResponse(302);
+
+        if (is_callable($this->failureDirectUrl)) {
+            return $this->redirectUrl($request, $response, $authResult, $authenticator);
+        }
+
+        return $response->withHeader('Location', $this->redirectUrl);
     }
 
     /**
-     * Callable implementation for the middleware stack.
+     * Process an incoming server request
      *
-     * @param \Psr\Http\Message\ServerRequestInterface $request The request.
-     * @param \Psr\Http\Message\ResponseInterface $response The response.
-     * @param callable $next The next middleware to call.
-     * @return \Psr\Http\Message\ResponseInterface A response.
+     * Processes an incoming server request in order to produce a response.
+     * If unable to produce the response itself, it may delegate to the provided
+     * request handler to do so.
      */
-    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, $next)
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
-        $service = $this->getAuthenticationService($request, $response);
+        $wasAuthenticated = $this->service->authenticate($request);
 
-        try {
-            $result = $service->authenticate($request, $response);
-        } catch (UnauthorizedException $e) {
-            $body = new Stream('php://memory', 'rw');
-            $body->write($e->getBody());
-            $response = $response->withStatus($e->getCode())
-                ->withBody($body);
-            foreach ($e->getHeaders() as $header => $value) {
-                $response = $response->withHeader($header, $value);
+        $authResult = $this->service->getResult();
+        $authenticator = $this->service->getSuccessfulAuthenticator();
+        $request = $request->withAttribute('authentication', $authResult);
+
+        if (!$wasAuthenticated) {
+            if (!empty($this->responseFactory) && !empty($this->unauthorizedRedirectUrl)) {
+                return $this->getUnauthorizedRedirectResponse($request, $authResult, $authenticator);
             }
 
-            return $response;
+            return $handler->handle($request);
         }
 
-        $request = $result['request'];
-        $request = $request->withAttribute($this->getConfig('identityAttribute'), $service->getIdentity());
-        $request = $request->withAttribute('authentication', $service);
-        $request = $request->withAttribute('authenticationResult', $result['result']);
+        if (!empty($this->responseFactory) && !empty($this->successRedirectUrl)) {
+            $response = $this->getSuccessRedirectResponse($request, $authResult, $authenticator);
+            $result = $this->service->persistIdentity($request, $response);
 
-        $response = $result['response'];
-
-        try {
-            return $next($request, $response);
-        } catch (UnauthenticatedException $e) {
-            $target = $this->getConfig('unauthenticatedRedirect');
-            if ($target) {
-                $url = $this->getRedirectUrl($target, $request);
-
-                return $response
-                    ->withStatus(301)
-                    ->withHeader('Location', $url);
-            }
-            throw $e;
+            return $response['response'];
         }
+
+        $response = $handler->handle($request);
+        if ($response instanceof ResponseInterface) {
+            $result = $this->service->persistIdentity($request, $response);
+
+            return $result['response'];
+        }
+
+        return $response;
+    }
+
+    protected function getRedirectResponse(string $targetUrl): ResponseInterface
+    {
+        return $this->responseFactory
+                ->createResponse(302)
+                ->withHeader('Location', $targetUrl);
     }
 
     /**
@@ -140,7 +153,7 @@ class AuthenticationMiddleware
      * @param \Psr\Http\Message\ServerRequestInterface $request Request instance.
      * @return string
      */
-    protected function getRedirectUrl($target, ServerRequestInterface $request)
+    protected function getRedirectUrl(string $target, ServerRequestInterface $request): string
     {
         $param = $this->getConfig('queryParam');
         if ($param === null) {
@@ -148,7 +161,6 @@ class AuthenticationMiddleware
         }
 
         $query = urlencode($param) . '=' . urlencode($request->getUri());
-
         if (strpos($target, '?') !== false) {
             $query = '&' . $query;
         } else {
@@ -157,31 +169,4 @@ class AuthenticationMiddleware
 
         return $target . $query;
     }
-
-    /**
-     * Returns AuthenticationServiceInterface instance.
-     *
-     * @param \Psr\Http\Message\ServerRequestInterface $request Server request.
-     * @param \Psr\Http\Message\ResponseInterface $response Response.
-     * @return \Authentication\AuthenticationServiceInterface
-     * @throws \RuntimeException When authentication method has not been defined.
-     */
-    protected function getAuthenticationService($request, $response)
-    {
-        $subject = $this->subject;
-
-        if ($subject instanceof AuthenticationServiceProviderInterface) {
-            $subject = $this->subject->getAuthenticationService($request, $response);
-        }
-
-        if (!$subject instanceof AuthenticationServiceInterface) {
-            $type = is_object($subject) ? get_class($subject) : gettype($subject);
-            $message = sprintf('Service provided by a subject must be an instance of `%s`, `%s` given.', AuthenticationServiceInterface::class, $type);
-
-            throw new RuntimeException($message);
-        }
-
-        return $subject;
-    }
-
 }
